backend/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
cryptography==41.0.4
Django==4.2.5
django-appconf==1.0.5
django-cors-headers==4.2.0
django-filter==23.2
djangorestframework==3.14.0
idna==3.4
Markdown==3.4.4
pycparser==2.21
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
stripe==7.1.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.7
